Move all attestation_data fns into same file
